requirements.txt
Adafruit-Blinka==8.20.1
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-dht==4.0.2
adafruit-circuitpython-requests==2.0.1
adafruit-circuitpython-typing==1.9.4
Adafruit-PlatformDetect==3.49.0
Adafruit-PureIO==1.1.11
async-timeout==4.0.3
pyftdi==0.55.0
pyserial==3.5
pyusb==1.2.1
PyYAML==6.0.1
redis==5.0.0
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
sysv-ipc==1.1.0
typing-extensions==4.7.1
